Add Login Activity and also onboarding logic.
