+++ app/src/main/java/com/qibla/qiblacompass/prayertime/finddirection/views/login/LoginActivity.kt
@@ -1,0 +1,12 @@
+package com.qibla.qiblacompass.prayertime.finddirection.views.login
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.qibla.qiblacompass.prayertime.finddirection.R
+
+class LoginActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+    }
+}
