+++ app/src/main/res/drawable/indicator3.xml
@@ -1,0 +1,9 @@
+<vector android:autoMirrored="true" android:height="10dp"
+    android:viewportHeight="10" android:viewportWidth="38"
+    android:width="38dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillAlpha="0.25" android:fillColor="#2D6D5F"
+        android:pathData="M5,5m-5,0a5,5 0,1 1,10 0a5,5 0,1 1,-10 0" android:strokeAlpha="0.25"/>
+    <path android:fillAlpha="0.25" android:fillColor="#2D6D5F"
+        android:pathData="M19,5m-5,0a5,5 0,1 1,10 0a5,5 0,1 1,-10 0" android:strokeAlpha="0.25"/>
+    <path android:fillColor="#2D6D5F" android:pathData="M33,5m-5,0a5,5 0,1 1,10 0a5,5 0,1 1,-10 0"/>
+</vector>
