+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,7 @@
     <string name="qibla_compass">Qibla Compass</string>
     <string name="skip">Skip</string>
     <string name="next">Next</string>
+    <string name="test">Lorem ipsum dolor sit amet, consectetur adipiscing elit</string>
+    <string name="prayer_timing">Prayer Timing</string>
+    <string name="holy_quran">Holy Quran</string>
 </resources>
